AND-17999 Update duration to Duration type

+++ domain/src/main/kotlin/mega/privacy/android/domain/usecase/chat/message/CreateCallEndedMessageUseCase.kt
@@ -14,7 +14,7 @@ internal class CreateCallEndedMessageUseCase @Inject constructor() : CreateTyped
                 isMine = isMine,
                 userHandle = userHandle,
                 termCode = termCode,
-                duration = duration.toLong(),
+                duration = duration,
                 shouldShowAvatar = shouldShowAvatar,
                 shouldShowTime = shouldShowTime,
                 shouldShowDate = shouldShowDate,
